src/com/github/lucasefdr/B10TDD/service/ReajusteService.java
package com.github.lucasefdr.B10TDD.service;

import com.github.lucasefdr.B10TDD.model.Desempenho;
import com.github.lucasefdr.B10TDD.model.Funcionario;

import java.math.BigDecimal;
import java.math.RoundingMode;

public class ReajusteService {
    public void concederReajuste(Funcionario funcionario, Desempenho desempenho) {
        BigDecimal percentual = desempenho.percentualReajuste();
        BigDecimal reajuste = funcionario.getSalario().multiply(percentual);
        funcionario.reajustarSalario(reajuste);
    }
}
